+++ client/src/game/systems/settings/players/models.ts
@@ -6,6 +6,7 @@ export interface PlayerOptions {
     fowColour: string;
     rulerColour: string;
     useToolIcons: boolean;
+    showTokenDirections: boolean;
 
     // Behaviour
     invertAlt: boolean;
@@ -33,6 +34,7 @@ export interface ServerPlayerOptions {
     fow_colour: string;
     ruler_colour: string;
     use_tool_icons: boolean;
+    show_token_directions: boolean;
 
     invert_alt: boolean;
     disable_scroll_to_zoom: boolean;