+++ client/src/game/systems/settings/players/state.ts
@@ -17,6 +17,7 @@ const state = buildState<PlayerState, "gridSize">({
     fowColour: init("rgba(0, 0, 0, 1)"),
     rulerColour: init("rgba(255, 0, 0, 1)"),
     useToolIcons: init(true),
+    showTokenDirections: init(true),
 
     invertAlt: init(false),
     disableScrollToZoom: init(false),